// functions/src/types.ts
// Copyright 2019 Google LLC
//
// Licensed under the Apache License, Version 2.0 (the "License");
// you may not use this file except in compliance with the License.
// You may obtain a copy of the License at
//
//      http://www.apache.org/licenses/LICENSE-2.0
//
// Unless required by applicable law or agreed to in writing, software
// distributed under the License is distributed on an "AS IS" BASIS,
// WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
// See the License for the specific language governing permissions and
// limitations under the License.

/**
 * Interface representing the operation resource in AutoML Api
 *
 * https://cloud.google.com/vision/automl/alpha/docs/reference/rest/v1beta1/projects.locations.operations#Operation
 */
export interface OperationMetadata {
  name: string; // a name to identify operation by
  done: boolean; // whether the operation has been completed
  metadata: any;
}